migration fix

+++ database/migrations/2024_08_05_075008_create_tupals_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
                 ->on('players')
                 ->onUpdate('cascade')
                 ->onDelete('cascade');
-            $table->float('point');
+            $table->integer('point');
             $table->integer('reject');
             $table->integer('serve');
             $table->timestamps();
